Create state and density functions and eyekron
Renamed statezero to densityzero since it returns a density matrix, not a state vector.
Added statezero, which returns state vector of a system in state zero.
Added eyekron, that lets you calculate kron product of a funmat with identities on its left and right more efficiently than using kron with identities.